Project update 2022-10-21
Re-entrant cylindrical cavity with one post codeposited with Pd and D, quenched with liquid nitrogen, assembled in LN, drained, evacuated, back-filled with low-pressure D2, cold plasma activated with RF energy, tuned to an overtone of the Deuterium dipole resonance in the nuclear-active sites, which are thought to be deuterium dipoles resonating collectively in a Bose condensation in zones of super-abundant vacancies. The "sweet spot" might be the 27.3843525222 MHz hyperfine spin-spin transition in Deuterium. Thereby the fusion energy will be extracted directly as electricity.